+++ src/Products/MeetingCommunes/tests/testCustomToolPloneMeeting.py
@@ -5,7 +5,7 @@
 # GNU General Public License (GPL)
 #
 
-from plone.app.textfield.value import RichTextValue
+from imio.helpers.content import richtextval
 from Products.MeetingCommunes.tests.MeetingCommunesTestCase import MeetingCommunesTestCase
 
 
@@ -18,7 +18,7 @@ class testCustomToolPloneMeeting(MeetingCommunesTestCase):
         """
         self.changeUser('pmManager')
         m1 = self._createMeetingWithItems()
-        m1.assembly = RichTextValue(
+        m1.assembly = richtextval(
             'Pierre Dupont - Bourgmestre,\n'
             'Charles Exemple - 1er Echevin,\n'
             'Echevin Un, Echevin Deux, Echevin Trois - Echevins,\n'
@@ -32,7 +32,7 @@ class testCustomToolPloneMeeting(MeetingCommunesTestCase):
                          '')
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Excus'),
                          '')
-        m1.assembly = RichTextValue(
+        m1.assembly = richtextval(
             'Pierre Dupont - Bourgmestre,\n'
             'Charles Exemple - 1er Echevin,\n'
             'Echevin Un, Echevin Deux, Echevin Trois - Echevins,\n'
@@ -42,12 +42,12 @@ class testCustomToolPloneMeeting(MeetingCommunesTestCase):
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='')[0],
                          attendee)
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Absent'),
-                         '')
+                         u'')
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Excus')[0],
-                         'Excusés:')
+                         u'Excusés:')
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Excus')[1],
-                         '<p class="mltAssembly">Monsieur x, Mesdames Y et Z</p>')
-        m1.assembly = RichTextValue(
+                         u'<p class="mltAssembly">Monsieur x, Mesdames Y et Z</p>')
+        m1.assembly = richtextval(
             'Pierre Dupont - Bourgmestre,\n'
             'Charles Exemple - 1er Echevin,\n'
             'Echevin Un, Echevin Deux, Echevin Trois - Echevins,\n'
@@ -59,10 +59,10 @@ class testCustomToolPloneMeeting(MeetingCommunesTestCase):
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='')[0],
                          attendee)
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Absent')[0],
-                         'Absent:')
+                         u'Absent:')
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Absent')[1],
-                         '<p class="mltAssembly">Monsieur tartenpion</p>')
+                         u'<p class="mltAssembly">Monsieur tartenpion</p>')
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Excus')[0],
-                         'Excusés:')
+                         u'Excusés:')
         self.assertEqual(self.tool.adapted().getSpecificAssemblyFor(m1.get_assembly(), startTxt='Excus')[1],
-                         '<p class="mltAssembly">Monsieur x, Mesdames Y et Z</p>')
+                         u'<p class="mltAssembly">Monsieur x, Mesdames Y et Z</p>')