+++ src/Products/MeetingCommunes/tests/testCustomWorkflows.py
@@ -5,9 +5,11 @@
 # GNU General Public License (GPL)
 #
 
+from datetime import timedelta
 from Products.CMFCore.permissions import AccessContentsInformation
 from Products.CMFCore.permissions import View
 from Products.MeetingCommunes.tests.MeetingCommunesTestCase import MeetingCommunesTestCase
+from Products.PloneMeeting.utils import get_advice_alive_states
 
 
 class testCustomWorkflows(MeetingCommunesTestCase):
@@ -140,3 +142,91 @@ class testCustomWorkflows(MeetingCommunesTestCase):
         # we check that item and meeting did their complete workflow
         self.assertEqual(item.query_state(), 'accepted')
         self.assertEqual(meeting.query_state(), 'closed')
+
+    def test_pm_AdviceCustomWorkflowDelayAwareCorrectlyUpdated(self):
+        """When an delay aware advice using a custom workflow is in a custom review_state,
+           it is correctly updated by the night task."""
+        self._configureFinancesAdvice()
+        cfg = self.meetingConfig
+        cfg.setItemAdviceStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceEditStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceViewStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setCustomAdvisers((
+            {'row_id': 'unique_id_001',
+             'org': self.vendors_uid,
+             'for_item_created_from': '2024/01/01',
+             'delay': '10',
+             'delay_left_alert': '4',
+             'delay_label': 'Finance advice',
+             'is_linked_to_previous_row': '0'}, ))
+        # create item and ask 2 advices
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem',
+                           title="Item to advice",
+                           category='development',
+                           optionalAdvisers=('%s__rowid__unique_id_001' % self.vendors_uid, ))
+        self.changeUser('pmReviewer2')
+        vendors_advice_portal_type = item.adapted()._advicePortalTypeForAdviser(self.vendors_uid)
+        cfg.setDefaultAdviceHiddenDuringRedaction([vendors_advice_portal_type])
+        vendors_advice = self.addAdvice(
+            item,
+            **{'advice_group': self.vendors_uid,
+               'advice_type': u'positive_with_remarks',
+               'advice_comment': u'My comment',
+               'advice_portal_type': vendors_advice_portal_type})
+        # is delay aware
+        self.assertEqual(item.adviceIndex[self.vendors_uid]['delay'], '10')
+        # is in a custom review_state
+        self.assertEqual(vendors_advice.query_state(), 'proposed_to_financial_manager')
+        # is returned taken into account by @@update-delay-aware-advices query
+        self.changeUser('siteadmin')
+        query = self.portal.restrictedTraverse('@@update-delay-aware-advices')._computeQuery()
+        self.assertTrue(item.UID() in [brain.UID for brain in self.catalog(**query)])
+        # every not ended states are taken into account
+        self.assertEqual(
+            sorted(get_advice_alive_states()),
+            ['advice_under_edit', 'financial_advice_signed', 'proposed_to_financial_manager'])
+        # check that @@pm-night-tasks does the job
+        # avoid failing test due to delay computation
+        self.tool.setHolidays([])
+        self.tool.setDelayUnavailableEndDays([])
+        self.tool.setWorkingDays(('mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun', ))
+        self.tool.notifyModified()
+        # for now 10 days left
+        self.assertEqual(item.adviceIndex[self.vendors_uid]['delay_infos']['left_delay'], 10)
+        item.adviceIndex[self.vendors_uid]['delay_started_on'] = \
+                item.adviceIndex[self.vendors_uid]['delay_started_on'] - timedelta(1)
+        # after update by @@pm-night-tasks, 9 days left
+        self.portal.restrictedTraverse('@@pm-night-tasks')()
+        self.assertEqual(item.adviceIndex[self.vendors_uid]['delay_infos']['left_delay'], 9)
+
+    def test_pm_Show_advice_on_final_wf_transition_when_item_in_advice_not_giveable_state(self):
+        """Test especially that if a finances advice is taken back in a state
+           where it is no more giveable, it is not shown if advice WF was not ended."""
+        self._configureFinancesAdvice()
+        cfg = self.meetingConfig
+        cfg.setItemAdviceStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceEditStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceViewStates((self._stateMappingFor('itemcreated'), ))
+        # create item and ask advice
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem',
+                           title="Item to advice",
+                           category='development',
+                           optionalAdvisers=(self.vendors_uid, ))
+        self.changeUser('pmReviewer2')
+        vendors_advice_portal_type = item.adapted()._advicePortalTypeForAdviser(self.vendors_uid)
+        cfg.setDefaultAdviceHiddenDuringRedaction([vendors_advice_portal_type])
+        vendors_advice = self.addAdvice(
+            item,
+            **{'advice_group': self.vendors_uid,
+               'advice_type': u'positive_with_remarks',
+               'advice_comment': u'My comment',
+               'advice_portal_type': vendors_advice_portal_type})
+        self.assertEqual(vendors_advice.query_state(), 'proposed_to_financial_manager')
+        self.assertTrue(vendors_advice.advice_hide_during_redaction)
+        # propose item, it will still be hidden during redaction
+        self.proposeItem(item)
+        self.assertEqual(vendors_advice.query_state(), 'advice_given')
+        self.assertTrue(vendors_advice.advice_hide_during_redaction)
+        self.assertTrue(item.adviceIndex[self.vendors_uid]['hidden_during_redaction'])